支持通过配置文件选择RPC中使用的序列化器

+++ example-common/src/main/java/cn/edu/szu/example/common/service/UserService.java
@@ -10,7 +10,7 @@ public interface UserService {
      */
     User getUser(User user);
 
-    default float getNumber() {
-        return 10;
+    default double getNumber() {
+        return 10.0;
     }
 }
